requirements.txt
pandas
numpy
scikit-learn
flask
joblib
python-dotenv
google-generativeai
